% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{prediction with top ten features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{an n x 10 matric containing n new observations}
}
\value{
a numeric vector containing the predicted values
}
\description{
This function takes a set of coefficients produced by the \code{topten} function
and makes a prediction for each of the values provided in the input X matrix
}
