% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test}
\alias{test}
\title{do a test}
\usage{
test(x)
}
\arguments{
\item{x}{int}
}
\value{
a vector of random numbers
}
\description{
lalalalalala
}
\details{
lalalalalalalallalalalala
}
\author{
Shang Ju
}

