\name{topten-package}
\alias{topten-package}
\alias{topten}
\docType{package}
\title{
\packageTitle{topten}
}
\description{
\packageDescription{topten}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{topten}
\packageIndices{topten}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{topten}

Maintainer: \packageMaintainer{topten}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
