% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a Model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vectgor of coefficients from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on the top 10 features
in 'x' that are most predictive of 'y'
}
\details{
This funtion runs a univariate regression of y on each predictor in x and
calculates a p-value indicating the significance of the association
}
\seealso{
\code{lm}
}
\author{
amarquezp
}
