% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{Prediction with Top Ten features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n by 10 matrix containing n new observations}

\item{b}{a vector of coefs obtained from the \code{topten} function}
}
\value{
a numeric vector containing predicted values
}
\description{
This function takes a set of coefficients produced by the \code{topten}
function and makes a prediction for each of the values provided in the input
'X' matrix.
}
