% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{Prediction with topten features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing n new observations}

\item{b}{a vector of coefficients obtinad from the \code{topten} function}
}
\value{
a numeric vector containing the predicted values
}
\description{
This function takes a set of coeeficients produced by the \code{topten}
function and makes a prediction for each of the values provided in the input x matrix
}

