% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{Prediction with Top 10 features
This function takes a set of coefficients produced by the \code{topten}\ function
and makes a prediction for each of the values provided in the input 'X' matrix.}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{n*10 matrix containing n new observations}

\item{b}{vector of coefficients obtained from the \code{topten}\ function}
}
\value{
a numeric vectore containing the predicted values
}
\description{
Prediction with Top 10 features
This function takes a set of coefficients produced by the \code{topten}\ function
and makes a prediction for each of the values provided in the input 'X' matrix.
}

