% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{Prediction with Top Ten Features}
\usage{
predict10(predictors, coefs)
}
\arguments{
\item{predictors}{a n by 10 matrix containing n new observation}

\item{coefs}{a vector of coefficients obtained from the \code{topten} function}
}
\value{
a numeric vector containing the predicted values
}
\description{
This function take a set of coefficients produced by the \code{topten}
function and ameks a prediction for each fo the values provided in the
input predictors matrix.
}
