% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{predict10}
\alias{predict10}
\title{Preditcion with Top 10 features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing new observations}
}
\value{
a numeric vector containing the predicted values
}
