% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a Model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n by p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefs from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on the top 10 features
in "x" that are most predictive of "y"
}
\details{
This function runs a univariate regression of y on each predictor in x and
calculates a p-value indicating the significance of the association. The
final set of 10 predictors is taken from the 10 smallest p values
}
\seealso{
\code{lm}
}
\author{
Benedict Neo
}
