% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Buiding a model with topten features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on the top 10 features
in 'x' that are most predictive of 'y'.
}
\details{
This function runs a univariate regression of y on each predictors in x and calculates p-values indicating the significance of the association
the final set of 10 predicotors is taken from the 10 smallest p-values
}
\author{
Borye
}
\seealso{
\code{lm}
}

