% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a model with top 10 predictors}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of the n observations a p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model using the top 10 predictors
}
\description{
This function develops a linear predition model using top 10 featers in 'x' that are most predictive of 'y'
}
\details{
This function runs a univariate regression of y on each of the predictors in x and calculates a p-value 
indicating the significance of the association.  The final set of 10 predictors is taken from the 10 smallest p-values
}
\author{
costab06
}
\seealso{
\code{lm}
}

