% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{topten}
\alias{topten}
\title{Building a Model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a matrix of n observations of p variables}

\item{y}{a vector of n representing the response}
}
\value{
a vector of coefficients from the final fitted model
}
\description{
This function develops an algorithm based on the top ten features in 'x' that are most predictive of 'y'
}
\details{
This function runs a univariate regression of y on each predictor in x and calculates a p-value indicating the significance of the association. The final set of 10 predictors is taken from the smallest 10 p-values.
}
\seealso{
\code{lm}
}
\author{
Giulio Mario Martena
}
