% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficeints from teh final fitted mole with top 10 features
}
\description{
This function develops a prediction algorithm base on the top 10 features
in 'x' that are most predictive of 'y'.
}
\details{
This function runs a univariate regression of y on each predictor in x and
calculates a p-value indicating the significance of hte association. the final set of 10 predictors
is taken from the 10 smallest p-values
}
\author{
Aleksey Kramer
}
\seealso{
\code{lm}
}

