% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a model with top ten features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on the top ten features in
'x' that are most predictive of 'y'
}
\details{
This function runs a univariate regression on each predictor
}
\seealso{
\code{lm}
}
\author{
Copied from Roger Lpeng
}
