% Generated by roxygen2 (4.0.2): do not edit by hand
\name{topten}
\alias{topten}
\title{Building a Model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm based on the top the features
in 'x' that are most predictive of 'y
}
\details{
This function runs a univariate regression model of y on each predictor in x
and calculates ...
}
\author{
Piegiorgio Palla
}
\seealso{
\code{lm}
}

